fix icon to new component
